// apps/venture/src/plugins/store/ui/generated-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = {
	[K in keyof T]: T[K];
};
// ...
export type Query = {
	__typename?: 'Query';
	administrators: AdministratorList;
	administrator?: Maybe<Administrator>;
	activeAdministrator?: Maybe<Administrator>;
	/** Get a list of Assets */
	assets: AssetList;
	/** Get a single Asset by id */
	asset?: Maybe<Asset>;
	me?: Maybe<CurrentUser>;
	channels: Array<Channel>;
	channel?: Maybe<Channel>;
	activeChannel: Channel;
	collections: CollectionList;
	/** Get a Collection either by id or slug. If neither id nor slug is speicified, an error will result. */
	collection?: Maybe<Collection>;
	collectionFilters: Array<ConfigurableOperationDefinition>;
	countries: CountryList;
	country?: Maybe<Country>;
	customerGroups: CustomerGroupList;
	customerGroup?: Maybe<CustomerGroup>;
	customers: CustomerList;
	customer?: Maybe<Customer>;
	facets: FacetList;
	facet?: Maybe<Facet>;
	globalSettings: GlobalSettings;
	job?: Maybe<Job>;
	jobs: JobList;
	jobsById: Array<Job>;
	jobQueues: Array<JobQueue>;
	order?: Maybe<Order>;
	orders: OrderList;
	paymentMethods: PaymentMethodList;
	paymentMethod?: Maybe<PaymentMethod>;
	paymentMethodEligibilityCheckers: Array<ConfigurableOperationDefinition>;
	paymentMethodHandlers: Array<ConfigurableOperationDefinition>;
	productOptionGroups: Array<ProductOptionGroup>;
	productOptionGroup?: Maybe<ProductOptionGroup>;
	search: SearchResponse;
	/** List Products */
	products: ProductList;
	/** Get a Product either by id or slug. If neither id nor slug is speicified, an error will result. */
	product?: Maybe<Product>;
	/** List ProductVariants either all or for the specific product. */
	productVariants: ProductVariantList;
	/** Get a ProductVariant by id */
	productVariant?: Maybe<ProductVariant>;
	promotion?: Maybe<Promotion>;
	promotions: PromotionList;
	promotionConditions: Array<ConfigurableOperationDefinition>;
	promotionActions: Array<ConfigurableOperationDefinition>;
	roles: RoleList;
	role?: Maybe<Role>;
	shippingMethods: ShippingMethodList;
	shippingMethod?: Maybe<ShippingMethod>;
	shippingEligibilityCheckers: Array<ConfigurableOperationDefinition>;
	shippingCalculators: Array<ConfigurableOperationDefinition>;
	fulfillmentHandlers: Array<ConfigurableOperationDefinition>;
	testShippingMethod: TestShippingMethodResult;
	testEligibleShippingMethods: Array<ShippingMethodQuote>;
	tag: Tag;
	tags: TagList;
	taxCategories: Array<TaxCategory>;
	taxCategory?: Maybe<TaxCategory>;
	taxRates: TaxRateList;
	taxRate?: Maybe<TaxRate>;
	zones: Array<Zone>;
	zone?: Maybe<Zone>;
	stores: StoreList;
	store?: Maybe<Store>;
};
// ...
export type Administrator = Node & {
	__typename?: 'Administrator';
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	firstName: Scalars['String'];
	lastName: Scalars['String'];
	emailAddress: Scalars['String'];
	user: User;
	customFields?: Maybe<Scalars['JSON']>;
};

export type AdministratorList = PaginatedList & {
	__typename?: 'AdministratorList';
	items: Array<Administrator>;
	totalItems: Scalars['Int'];
};

export type Asset = Node & {
	__typename?: 'Asset';
	tags: Array<Tag>;
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	name: Scalars['String'];
	type: AssetType;
	fileSize: Scalars['Int'];
	mimeType: Scalars['String'];
	width: Scalars['Int'];
	height: Scalars['Int'];
	source: Scalars['String'];
	preview: Scalars['String'];
	focalPoint?: Maybe<Coordinate>;
	customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type Collection = Node & {
	__typename?: 'Collection';
	isPrivate: Scalars['Boolean'];
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	languageCode?: Maybe<LanguageCode>;
	name: Scalars['String'];
	slug: Scalars['String'];
	breadcrumbs: Array<CollectionBreadcrumb>;
	position: Scalars['Int'];
	description: Scalars['String'];
	featuredAsset?: Maybe<Asset>;
	assets: Array<Asset>;
	parent?: Maybe<Collection>;
	children?: Maybe<Array<Collection>>;
	filters: Array<ConfigurableOperation>;
	translations: Array<CollectionTranslation>;
	productVariants: ProductVariantList;
	customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type Customer = Node & {
	__typename?: 'Customer';
	groups: Array<CustomerGroup>;
	history: HistoryEntryList;
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	title?: Maybe<Scalars['String']>;
	firstName: Scalars['String'];
	lastName: Scalars['String'];
	phoneNumber?: Maybe<Scalars['String']>;
	emailAddress: Scalars['String'];
	addresses?: Maybe<Array<Address>>;
	orders: OrderList;
	user?: Maybe<User>;
	customFields?: Maybe<Scalars['JSON']>;
};

export type CustomerHistoryArgs = {
	options?: Maybe<HistoryEntryListOptions>;
};

export type CustomerOrdersArgs = {
	options?: Maybe<OrderListOptions>;
};

export type CustomerGroupList = PaginatedList & {
	__typename?: 'CustomerGroupList';
	items: Array<CustomerGroup>;
	totalItems: Scalars['Int'];
};
// ...
export type Facet = Node & {
	__typename?: 'Facet';
	isPrivate: Scalars['Boolean'];
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	languageCode: LanguageCode;
	name: Scalars['String'];
	code: Scalars['String'];
	values: Array<FacetValue>;
	translations: Array<FacetTranslation>;
	customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type GlobalSettings = {
	__typename?: 'GlobalSettings';
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	availableLanguages: Array<LanguageCode>;
	trackInventory: Scalars['Boolean'];
	outOfStockThreshold: Scalars['Int'];
	serverConfig: ServerConfig;
	customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type JobList = PaginatedList & {
	__typename?: 'JobList';
	items: Array<Job>;
	totalItems: Scalars['Int'];
};

export type Job = Node & {
	__typename?: 'Job';
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	startedAt?: Maybe<Scalars['DateTime']>;
	settledAt?: Maybe<Scalars['DateTime']>;
	queueName: Scalars['String'];
	state: JobState;
	progress: Scalars['Float'];
	data?: Maybe<Scalars['JSON']>;
	result?: Maybe<Scalars['JSON']>;
	error?: Maybe<Scalars['JSON']>;
	isSettled: Scalars['Boolean'];
	duration: Scalars['Int'];
};

export type JobQueue = {
	__typename?: 'JobQueue';
	name: Scalars['String'];
	running: Scalars['Boolean'];
};

export type Order = Node & {
	__typename?: 'Order';
	nextStates: Array<Scalars['String']>;
	modifications: Array<OrderModification>;
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	/**
	 * The date & time that the Order was placed, i.e. the Customer
	 * completed the checkout and the Order is no longer "active"
	 */
	orderPlacedAt?: Maybe<Scalars['DateTime']>;
	/** A unique code for the Order */
	code: Scalars['String'];
	state: Scalars['String'];
	/** An order is active as long as the payment process has not been completed */
	active: Scalars['Boolean'];
	customer?: Maybe<Customer>;
	shippingAddress?: Maybe<OrderAddress>;
	billingAddress?: Maybe<OrderAddress>;
	lines: Array<OrderLine>;
	/**
	 * Surcharges are arbitrary modifications to the Order total which are neither
	 * ProductVariants nor discounts resulting from applied Promotions. For store,
	 * one-off discounts based on customer interaction, or surcharges based on payment
	 * methods.
	 */
	surcharges: Array<Surcharge>;
	discounts: Array<Discount>;
	/** An array of all coupon codes applied to the Order */
	couponCodes: Array<Scalars['String']>;
	/** Promotions applied to the order. Only gets populated after the payment process has completed. */
	promotions: Array<Promotion>;
	payments?: Maybe<Array<Payment>>;
	fulfillments?: Maybe<Array<Fulfillment>>;
	totalQuantity: Scalars['Int'];
	/**
	 * The subTotal is the total of all OrderLines in the Order. This figure also includes any Order-level
	 * discounts which have been prorated (proportionally distributed) amongst the OrderItems.
	 * To get a total of all OrderLines which does not account for prorated discounts, use the
	 * sum of `OrderLine.discountedLinePrice` values.
	 */
	subTotal: Scalars['Int'];
	/** Same as subTotal, but inclusive of tax */
	subTotalWithTax: Scalars['Int'];
	currencyCode: CurrencyCode;
	shippingLines: Array<ShippingLine>;
	shipping: Scalars['Int'];
	shippingWithTax: Scalars['Int'];
	/** Equal to subTotal plus shipping */
	total: Scalars['Int'];
	/** The final payable amount. Equal to subTotalWithTax plus shippingWithTax */
	totalWithTax: Scalars['Int'];
	/** A summary of the taxes being applied to this Order */
	taxSummary: Array<OrderTaxSummary>;
	history: HistoryEntryList;
	customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type PaymentMethodList = PaginatedList & {
	__typename?: 'PaymentMethodList';
	items: Array<PaymentMethod>;
	totalItems: Scalars['Int'];
};
// ...
export type PaymentMethod = Node & {
	__typename?: 'PaymentMethod';
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	name: Scalars['String'];
	code: Scalars['String'];
	description: Scalars['String'];
	enabled: Scalars['Boolean'];
	checker?: Maybe<ConfigurableOperation>;
	handler: ConfigurableOperation;
};
// ...
export type AssetList = PaginatedList & {
	__typename?: 'AssetList';
	items: Array<Asset>;
	totalItems: Scalars['Int'];
};

export enum AssetType {
	Image = 'IMAGE',
	Video = 'VIDEO',
	Binary = 'BINARY',
}

export type CurrentUser = {
	__typename?: 'CurrentUser';
	id: Scalars['ID'];
	identifier: Scalars['String'];
	channels: Array<CurrentUserChannel>;
};

export type CurrentUserChannel = {
	__typename?: 'CurrentUserChannel';
	id: Scalars['ID'];
	token: Scalars['String'];
	code: Scalars['String'];
	permissions: Array<Permission>;
};

export type Channel = Node & {
	__typename?: 'Channel';
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	code: Scalars['String'];
	token: Scalars['String'];
	defaultTaxZone?: Maybe<Zone>;
	defaultShippingZone?: Maybe<Zone>;
	defaultLanguageCode: LanguageCode;
	currencyCode: CurrencyCode;
	pricesIncludeTax: Scalars['Boolean'];
	customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type CollectionList = PaginatedList & {
	__typename?: 'CollectionList';
	items: Array<Collection>;
	totalItems: Scalars['Int'];
};
// ...
export enum SortOrder {
	Asc = 'ASC',
	Desc = 'DESC',
}
// ...
export type ConfigurableOperationDefinition = {
	__typename?: 'ConfigurableOperationDefinition';
	code: Scalars['String'];
	args: Array<ConfigArgDefinition>;
	description: Scalars['String'];
};
// ...
export type Country = Node & {
	__typename?: 'Country';
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	languageCode: LanguageCode;
	code: Scalars['String'];
	name: Scalars['String'];
	enabled: Scalars['Boolean'];
	translations: Array<CountryTranslation>;
};

export type CountryTranslation = {
	__typename?: 'CountryTranslation';
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	languageCode: LanguageCode;
	name: Scalars['String'];
};

export type CountryList = PaginatedList & {
	__typename?: 'CountryList';
	items: Array<Country>;
	totalItems: Scalars['Int'];
};
// ...
export type CustomerGroup = Node & {
	__typename?: 'CustomerGroup';
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	name: Scalars['String'];
	customers: CustomerList;
};

export type CustomerGroupCustomersArgs = {
	options?: Maybe<CustomerListOptions>;
};

export type CustomerList = PaginatedList & {
	__typename?: 'CustomerList';
	items: Array<Customer>;
	totalItems: Scalars['Int'];
};
// ...
export type FacetList = PaginatedList & {
	__typename?: 'FacetList';
	items: Array<Facet>;
	totalItems: Scalars['Int'];
};
// ...
export type OrderList = PaginatedList & {
	__typename?: 'OrderList';
	items: Array<Order>;
	totalItems: Scalars['Int'];
};
// ...
export type StoreListOptions = {
	skip?: Maybe<Scalars['Int']>;
	take?: Maybe<Scalars['Int']>;
	sort?: Maybe<StoreSortParameter>;
	filter?: Maybe<StoreFilterParameter>;
};
// ...
	export type Variables = CreateStoreMutationVariables;
// ...
export namespace GetStores {
	export type Variables = GetStoresQueryVariables;
	export type Query = GetStoresQuery;
	export type Stores = NonNullable<GetStoresQuery['stores']>;
	export type Items = NonNullable<
		NonNullable<NonNullable<GetStoresQuery['stores']>['items']>[number]
	>;
}

export namespace GetStore {
	export type Variables = GetStoreQueryVariables;
	export type Query = GetStoreQuery;
	export type Store = NonNullable<GetStoreQuery['store']>;
}

export type CreateStoreMutationVariables = Exact<{
	input: CreateStoreInput;
}>;

export type CreateStoreMutation = { __typename?: 'Mutation' } & {
	createStore: { __typename?: 'Store' } & Pick<Store, 'id' | 'name'>;
};

export type UpdateStoreMutationVariables = Exact<{
	input: UpdateStoreInput;
}>;

export type UpdateStoreMutation = { __typename?: 'Mutation' } & {
	updateStore: { __typename?: 'Store' } & Pick<Store, 'id' | 'name'>;
};

export type GetStoresQueryVariables = Exact<{
	options?: Maybe<StoreListOptions>;
}>;

export type GetStoresQuery = { __typename?: 'Query' } & {
	stores: { __typename?: 'StoreList' } & Pick<StoreList, 'totalItems'> & {
			items: Array<
				{ __typename?: 'Store' } & Pick<
					Store,
					'id' | 'name' | 'createdAt'
				>
			>;
		};
};